# api/openapi.yaml
openapi: 3.1.0
info:
  title: Panda Cooking API
  version: 1.0.0
  summary: API de receitas do Panda Cooking.
  description: |
    API REST do Panda Cooking: cadastro e login, receitas com fotos,
    ingredientes e modo de preparo, comentários e favoritos.

    **Sessão.** O login devolve um access token (JWT, 15 minutos) no corpo e
    grava o refresh token no cookie `panda_refresh` (HttpOnly,
    SameSite=Strict, Path=/api/auth). Quando o access token vence, o cliente
    chama `POST /api/auth/refresh` e recebe outro. O refresh token troca a
    cada uso; usar de novo um token já trocado derruba a sessão.

    **Erros.** Todo erro tem o formato `{"error": "mensagem"}`. Erro de
    validação (422) traz também `fields`, com a mensagem de cada campo.

    **Limites.** Por IP: 300 requisições por minuto em /api, 10 tentativas de
    login por minuto, 30 renovações por minuto e 10 cadastros por hora.
    Depois de 5 senhas erradas, o login daquele e-mail fica bloqueado por 15
    minutos. Passou do limite, a resposta é 429 com `Retry-After`.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local (docker compose)
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Sessão]
      summary: Renovar o access token
      description: |
        Troca o refresh token do cookie por um novo e devolve outro access
        token. Token já usado é aceito por 20 segundos (duas abas renovando
        juntas); depois disso, usar de novo derruba a sessão.
      operationId: refresh
      security:
        - refreshCookie: []
      responses:
        "200":
          description: Sessão renovada.
          headers:
            Set-Cookie: { $ref: "#/components/headers/SessionCookies" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "401":
          description: Sessão vencida, encerrada ou inexistente. Os cookies são apagados.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/CrossOrigin" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token devolvido pelo login e pela renovação (vale 15 minutos).
    refreshCookie:
      type: apiKey
      in: cookie
      name: panda_refresh
      description: Refresh token gravado pelo login (HttpOnly, SameSite=Strict, Path=/api/auth).
  headers:
    SessionCookies:
      description: "`panda_refresh` (o refresh token) e `panda_session` (vale 1; só avisa o front que existe sessão)."
      schema: { type: string }
...
  responses:
...
    CrossOrigin:
      description: Requisição de navegador vinda de outro site (proteção contra CSRF).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: Passou do limite de requisições.
      headers:
        Retry-After:
          description: Segundos até poder tentar de novo.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: receita não encontrada }
...
    AuthResult:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, description: Validade do access token em segundos., example: 900 }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, name, email, image_profile, is_adm, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        image_profile: { type: string, description: Link da foto; vazio quando não tem. }
        is_adm: { type: boolean }
        created_at: { type: string, format: date-time }
...